move animation logic to popup
